==> notebooks/1_get_age_dependent_exon_genes-generate_psi_meta_dfs.ipynb <==
## This script make the psi matrix into right shape.
This script also performs filtering to remove QC-failed samples, and remove exons with too many missing PSIs

library(ggstatsplot)
library(ggside)
library(dplyr)
library(summarytools)
library(ggfortify)

dir <- "/Users/suzheng/Documents/suzheng/UNSW/UNSWTasks/2022/psi_calculation/results/genomewide_PSI/psi_quanti/target_gene_psi"
setwd(dir)

len_psi <- read.csv("all.target_gene_psi.pasted.added_gene_symbols", header=T, row.names=1, sep="\t", check.names=F)
#len_psi <- read.csv("all.target_gene_inc.pasted.added_gene_symbols", header=T, row.names=1, sep="\t", check.names=F)

qc_failed_samples <- read.csv("all_qc_failed_samples.txt",  header=F)

len <- len_psi$length
psi <- len_psi[, -1]

qc_passed_samples <- !(colnames(psi) %in% qc_failed_samples[,1])

psi <- psi[,qc_passed_samples]

na_frac <- apply(psi, 1, function(x){sum(is.na(x))/length(x)})

### Distrubition of the missing value frequency for exons

psi_NA_Fraction_by_gene_df <- data.frame(na_frac = na_frac)
options(repr.plot.width=7.5, repr.plot.height=7.5)
# Generate histogram
psi_NA_Fraction_by_gene <- ggplot(psi_NA_Fraction_by_gene_df, aes(x=na_frac))
save(psi_NA_Fraction_by_gene, file=paste0(Sys.getenv("psi_fig_tables_RData_dir"), "/psi_NA_Fraction_by_gene.RData"))

psi <- psi[na_frac<0.4, ]

dim(psi)

gtex_meta <- read.csv("/Users/suzheng/Documents/suzheng/UNSW/UNSWTasks/2021/velocityRNA/data/meta_data/sample_attributes.txt", header=T, sep="\t")

gtex_meta %>% filter(SAMPID  %in% colnames(psi)) %>% dim()

pw_meta <- read.csv("/Users/suzheng/Documents/suzheng/UNSW/UNSWTasks/2021/Broad_Perth_RNA_seq_titinopathy/meta/sample_ages_gender.txt", header=T, sep="\t")
gtex_meta <- read.csv("/Users/suzheng/Documents/suzheng/UNSW/UNSWTasks/2021/velocityRNA/data/meta_data/subject_phenotypes.txt", header=T, sep="\t")
rownames(pw_meta) <- pw_meta$sample
rownames(gtex_meta) <- gtex_meta$SUBJID

mcm_meta <- read.csv("/Users/suzheng/Documents/suzheng/UNSW/UNSWTasks/2022/psi_calculation/results/genomewide_PSI/meta/E-MTAB-6814.sdrf.txt.clean.added_months", header=T, sep="\t")



gtex_sample_meta <- read.csv(
"/Users/suzheng/Documents/suzheng/UNSW/UNSWTasks/2021/velocityRNA/data/meta_data/sample_attributes.txt",
    header=T,
    sep="\t"
)


gtex_sample_meta <- gtex_sample_meta[,c("SAMPID", "SMTS", "SMTSD")]
rownames(gtex_sample_meta) <- gtex_sample_meta$SAMPID

pw_index <- !grepl("GTEX", colnames(psi)) & !grepl("Human", colnames(psi))
gtex_index <- grepl("GTEX", colnames(psi))
mcm_index <- grepl("Human", colnames(psi))

pw_names <- intersect(colnames(psi)[pw_index], pw_meta$sample)
mcm_names <- intersect(colnames(psi)[mcm_index], mcm_meta$source.name)
psi_reordered <- cbind(psi[,c(pw_names,mcm_names)], psi[,gtex_index])
gtex_names <- colnames(psi[,gtex_index])

gtex_sample_meta_reordered <- gtex_sample_meta[gtex_names,]

row.names(mcm_meta) <- mcm_meta$source.name
pw_meta_reordered <- pw_meta[pw_names,]
mcm_meta_reordered <- mcm_meta[mcm_names,]

psi_gtex_subject_ids <- sapply(
    colnames(psi)[gtex_index],
    function(x)paste(strsplit(x, "-")[[1]][1], strsplit(x, "-")[[1]][2], sep="-")
    )
gtex_meta_reordered <- gtex_meta[match(psi_gtex_subject_ids, gtex_meta$SUBJID), ]


meta <- data.frame(month=c(pw_meta_reordered[, "month"], mcm_meta_reordered$month,12*gtex_meta_reordered[, "AGE"]),
                   gender=c(pw_meta_reordered[, "genderM1F2"], mcm_meta_reordered$sex, gtex_meta_reordered[, "SEX"]),
                   group=c(rep("Pathwest", length(pw_names)), rep("MCM", dim(mcm_meta_reordered)[1]), rep("GTEx", dim(gtex_meta_reordered)[1])),
                   tissue=c(rep("skeletal_muscle", length(pw_names)), mcm_meta_reordered$organism.part, gtex_sample_meta_reordered$SMTS),
                   subtissue=c(rep("skeletal_muscle", length(pw_names)), mcm_meta_reordered$organism.part, gtex_sample_meta_reordered$SMTSD),
                   disease=c(rep("normal", dim(psi_reordered)[2]))
                  )

meta$gender[meta$gender==1 | meta$gender=="male"] <- "Male"
meta$gender[meta$gender==2 | meta$gender=="female"] <- "Female"
rownames(meta) <- colnames(psi_reordered)
meta <- meta %>% mutate(age_group=ifelse(month < 1, "fetal_neonatal", ifelse(month < 18*12, "pediatric", "adult")))


meta$age_group <- factor(meta$age_group, levels=c("fetal_neonatal", "pediatric", "adult"))

meta[meta$tissue=="Heart", "tissue"] <- "heart"
meta[meta$tissue=="Muscle", "tissue"] <- "skeletal_muscle"
meta[meta$subtissue=="Muscle - Skeletal", "subtissue"] <- "skeletal_muscle"

Copied to overview_of_meta_data.html

save(psi, psi_reordered, meta, file = "/Users/suzheng/Documents/suzheng/UNSW/UNSWTasks/2022/psi_calculation/results/genomewide_PSI/psi_quanti/target_gene_psi/data/psi_meta.Rdata")

dim(psi)
dim(psi_reordered)
dim(meta)

table(meta[c("tissue", "age_group")])
