==> notebooks/5_get_age_dependent_exon_genes-assoc_analysis.ipynb <==
library(ggstatsplot)
library(ggside)
library(dplyr)
library(summarytools)
library(ggfortify)
library(plyr)
library(easyPubMed)

Copied to overview_of_meta_data.html

Sys.setenv(psi_fig_tables_RData_dir="/Users/suzheng/Documents/suzheng/UNSW/UNSWTasks/2022/psi_calculation/results/genomewide_PSI/psi_quanti/fig_tables/RData")

ANALYZE_ORIGINAL_DATA = FALSE
if(ANALYZE_ORIGINAL_DATA){
    setwd("/Users/suzheng/Documents/suzheng/UNSW/UNSWTasks/2022/psi_calculation/results/genomewide_PSI/psi_quanti/target_gene_psi/")
    load("/Users/suzheng/Documents/suzheng/UNSW/UNSWTasks/2022/psi_calculation/results/genomewide_PSI/psi_quanti/target_gene_psi/data/psi_meta.Rdata")
}else{
    setwd("/Users/suzheng/Documents/suzheng/UNSW/UNSWTasks/2022/psi_calculation/results/genomewide_PSI/psi_quanti/target_gene_psi_IRcorrected/")
    load("/Users/suzheng/Documents/suzheng/UNSW/UNSWTasks/2022/psi_calculation/results/genomewide_PSI/psi_quanti/target_gene_psi_IRcorrected/data/psi_meta.Rdata")
}



psi_var_fil <- t(psi_reordered)

factor_cmp <- function(col_of_interest, feature, p_threshold=0.001, dat=NA, plot=FALSE){
    p <- NA
    rho <- NA
    return(list(p=1.0, rho=0))
    if(var(dat[[col_of_interest]], na.rm =T) ==0){
        return(list(p=1.0, rho=0))
    }
    feature_number <- dat[[feature]][!is.na(dat[[col_of_interest]])] %>% unique() %>% length()
    if((is.character(dat[[feature]]) | is.factor(dat[[feature]])) & feature_number > 1){
        if (feature_number >2){
            p <- kruskal.test(dat[[col_of_interest]] ~ dat[[feature]])$p.value
        }else{
            p <- wilcox.test(dat[[col_of_interest]] ~ dat[[feature]])$p.value
        }
        if(plot & p < p_threshold){
            g <- ggbetweenstats(
              x     = !!rlang::sym(feature),
              y     = !!rlang::sym(col_of_interest),
              data = dat
            )
            print(g)
        }
    }
    if(is.numeric(dat[[feature]])){
        cor_obj <- cor.test(dat[[col_of_interest]], dat[[feature]], method = "spearman", exact = FALSE)
        p <- cor_obj$p.value
        rho <- cor_obj$estimate
        if(plot & p < p_threshold){
            g <- ggscatterstats(
              x     = !!rlang::sym(feature),
              y     = !!rlang::sym(col_of_interest),
              data = dat,
              type = "np" # "np" (nonparametric/spearman)
            )
            print(g)
        }
    }
    return(list(p=p, rho=rho))
    #return(p)
}

# selected_samples is boolean list
# grouping_col is string like 'age_group'
get_assoc_result_for_selected_samples <- function(selected_samples, grouping_col){
    selected_meta <- meta[selected_samples,]
    selected_psi <- psi_var_fil[selected_samples,]
    dim(selected_psi)
    dim(selected_meta)
    dat <- cbind(selected_meta, selected_psi)
    options(warn=-1)
    tissue_gtex_pvals <- sapply(
        colnames(selected_psi), 
        function(x)factor_cmp(x, grouping_col, p_threshold=0.05/dim(selected_psi)[2], dat=dat)
        )
    options(warn=0)
    assoc_pvals <- as.data.frame(t(tissue_gtex_pvals)[,1:2] )
    assoc_result <- data.frame(pval=unlist(assoc_pvals$p), name=rownames(assoc_pvals), estimate=unlist(assoc_pvals$rho))
    assoc_result$gene <- sapply(assoc_result$name, function(x)strsplit(strsplit(x, ":")[[1]][2], '%')[[1]][1])
    assoc_result$exon_num <- sapply(assoc_result$name, function(x)strsplit(x, ":")[[1]][1])
    assoc_result$fdr <- p.adjust(assoc_result$pval, method="BH") 
    assoc_result$signif <- assoc_result$fdr <0.05
    assoc_result
}

add_delta_mean <- function(assoc_result, dat, feature="age_group"){

    feature_means <- aggregate(dat[,assoc_result$name], by=list(tissue=dat[,feature]), FUN=function(x){mean(x, na.rm=T)})
    feature_sds <- aggregate(dat[,assoc_result$name], by=list(tissue=dat[,feature]), FUN=function(x){sd(x, na.rm=T)})

    assoc_result.1 <- cbind(assoc_result, t(feature_means)[-1,], t(feature_sds)[-1,])
    assoc_result.1.len <- dim(assoc_result.1)[2]
    mean_var_cols <- 8:assoc_result.1.len
    for(i in mean_var_cols){
        assoc_result.1[,i] <- as.numeric(assoc_result.1[,i])
    }
    colnames(assoc_result.1)[mean_var_cols] <- c(paste("mean", feature_means[,1], sep="."), paste("sd", feature_means[,1], sep="."))

    assoc_result.1$delta_mean <- max(abs(assoc_result.1[,8] - assoc_result.1[,9]), 
                                     abs(assoc_result.1[,8] - assoc_result.1[,10]),
                                     abs(assoc_result.1[,9] - assoc_result.1[,10])
                                    )

    delta1 <- abs(assoc_result.1[,8] - assoc_result.1[,9])
    delta2 <- abs(assoc_result.1[,8] - assoc_result.1[,10])
    delta3 <- abs(assoc_result.1[,9] - assoc_result.1[,10])

    assoc_result.1$delta_mean <- apply(cbind(delta1, delta2, delta3), 1, function(x){max(x)})
    assoc_result.sig <- assoc_result.1 %>% filter(signif == TRUE)
    #plot(assoc_result.sig$delta_mean, -log10(assoc_result.sig$fdr))
    #plot(density(assoc_result.sig %>% filter(!is.na(delta_mean)) %>% pull(delta_mean)))
    assoc_result.1
}

feature <- "age_group"
selected_samples <- meta$tissue == "skeletal_muscle" & meta$disease == "normal"


selected_meta <- meta[selected_samples,]
selected_psi <- psi_var_fil[selected_samples,]
dat <- cbind(selected_meta, selected_psi)


skeletal_muscle_age_group_assoc <- get_assoc_result_for_selected_samples(selected_samples, feature)

skeletal_muscle_age_group_assoc.1 <- add_delta_mean(skeletal_muscle_age_group_assoc, dat)

skeletal_muscle_age_group_assoc.1 <- skeletal_muscle_age_group_assoc.1 %>% mutate(gene0=sub("\\|\\w+", "", gene))

skeletal_muscle_age_group_assoc.1

write.table(skeletal_muscle_age_group_assoc.1, "skeletal_muscle_age_group_assoc.1.tsv", quote=F, sep="\t", col.names=NA)

## Below is for cardiac vs age groups

feature <- "age_group"
selected_samples <- meta$tissue == "heart" & meta$disease == "normal"
selected_meta <- meta[selected_samples,]
selected_psi <- psi_var_fil[selected_samples,]
dat <- cbind(selected_meta, selected_psi)


cardiac_muscle_age_group_assoc <- get_assoc_result_for_selected_samples(selected_samples, feature)


cardiac_muscle_age_group_assoc.1 <- add_delta_mean(cardiac_muscle_age_group_assoc, dat)

cardiac_muscle_age_group_assoc.1 <- cardiac_muscle_age_group_assoc.1 %>% mutate(gene0=sub("\\|\\w+", "", gene))

write.table(cardiac_muscle_age_group_assoc.1, "cardiac_muscle_age_group_assoc.1.tsv", quote=F, sep="\t", col.names=NA)

save(skeletal_muscle_age_group_assoc.1, 
     cardiac_muscle_age_group_assoc.1, 
     file="data/age_group.assoc.1.Rdata")
