==> notebooks/8_exon_usage_cor_clinvar_target_genes-cal_var_count_every_exon.ipynb <==
library(ggstatsplot)
library(ggside)
library(dplyr)
library(summarytools)
library(ggfortify)
library(plyr)
library(purrr)

dir <- "/Users/suzheng/Documents/suzheng/UNSW/UNSWTasks/2022/psi_calculation/results/genomewide_PSI/psi_quanti/cor_with_clinvar/anal/"
setwd(dir)

load("/Users/suzheng/Documents/suzheng/UNSW/UNSWTasks/2022/psi_calculation/results/genomewide_PSI/psi_quanti/target_gene_psi/data/psi_meta.Rdata")
psi_dir <- "/Users/suzheng/Documents/suzheng/UNSW/UNSWTasks/2022/psi_calculation/results/genomewide_PSI/psi_quanti/target_gene_psi/"


len_psi <- read.csv("../all.target_gene_psi.pasted.added_gene_symbols.clean", header=T, sep="\t", check.names=F)

gene_info <- len_psi[,1:4]
gene_info <- gene_info[match(rownames(psi_reordered), gene_info$exon),]

exon_coords <- read.csv("../gencode.v38.annotation.Exonic_Parts.LRG.noDup.gtf.exon_coord.target_genes", header=F, sep="\t")

colnames(exon_coords) <- c("chr", "start", "end", "exon")

gene_info.coord <- left_join(gene_info, exon_coords, by=c("exon0"="exon"))

gene_info.coord$index <- 1:dim(gene_info.coord)[1]

clinvar.rt <- read.csv("../clinvar.vcf.selected_genes.tsv", header=T, sep="\t")

clinvar.rt %>% dim()
psi %>%  dim()

clinvar_patho_types <- c("Pathogenic", "Likely_pathogenic", "Pathogenic/Likely_pathogenic", "Pathogenic|_other", "Pathogenic|_drug_response", "Likely_pathogenic|_drug_response", "Likely_pathogenic|_association", "Likely_pathogenic,_low_penetrance", "Likely_pathogenic|_risk_factor", "Pathogenic|_risk_factor", "Pathogenic/Likely_pathogenic|_other", "Pathogenic/Likely_pathogenic|_risk_factor", "Pathogenic|_association|_protective")

#gene_info.path_var_stat <- gene_info.coord[, c("chr", "start", "end")] %>% pmap_dfr(stat_path_var_in_one_exon)
#save(gene_info.path_var_stat, file = "/Users/suzheng/Documents/suzheng/UNSW/UNSWTasks/2022/psi_calculation/results/genomewide_PSI/psi_quanti/target_gene_psi/data/gene_info.path_var_stat.Rdata")

gnomad.var <- read.csv("/Users/suzheng/Documents/suzheng/UNSW/UNSWTasks/2022/psi_calculation/results/genomewide_PSI/psi_quanti/gnomAD_var_anal/all_genes.gnomAD_var.csv")

gnomad.var <- rename(gnomad.var, replace=c("Chromosome"="CHROM", "Position"="POS", "VEP.Annotation"="clin_sig"))

stat_some_types_of_var_in_every_exon <- function(var_df, target_var_types, res_col_name){
    stat_var_in_one_exon <- function(chr, start, end){
        vars.df <- var_df %>% filter(CHROM==chr & POS>=start & POS<=end)
        path.num <- sum(vars.df$clin_sig %in% target_var_types)
        df <- data.frame(path.num=path.num)
        colnames(df)[1] <- res_col_name
        df
    }

    gene_info.var_stat <- gene_info.coord[, c("chr", "start", "end")] %>% pmap_dfr(stat_var_in_one_exon)
    gene_info.var_stat
}

clinvar.path_var <-  stat_some_types_of_var_in_every_exon(clinvar.rt, clinvar_patho_types, "clinvar.path_var")
gnomad.missense <-  stat_some_types_of_var_in_every_exon(gnomad.var, c("missense_variant"), "gnomad.missense")
gnomad.syn <-  stat_some_types_of_var_in_every_exon(gnomad.var, c("synonymous_variant"), "gnomad.syn")
gnomad.frameshift <-  stat_some_types_of_var_in_every_exon(gnomad.var, c("frameshift_variant"), "gnomad.frameshift")
gnomad.missense_inframe <-  stat_some_types_of_var_in_every_exon(gnomad.var, c("missense_variant", "inframe_deletion", "inframe_insertion"), "gnomad.missense_inframe")
gnomad.frameshift_stopgain <-  stat_some_types_of_var_in_every_exon(gnomad.var, c("frameshift_variant", "stop_gained"), "gnomad.frameshift_stopgain")

gene_info.stat <- cbind(gene_info.coord, 
                        clinvar.path_var, 
                        gnomad.missense, 
                        gnomad.syn, 
                        gnomad.frameshift, 
                        gnomad.missense_inframe,
                        gnomad.frameshift_stopgain
                       )



gene_info.stat.file = "/Users/suzheng/Documents/suzheng/UNSW/UNSWTasks/2022/psi_calculation/results/genomewide_PSI/psi_quanti/target_gene_psi/data/gene_info.exon_var_stat.Rdata"
save(gene_info.stat, file = gene_info.stat.file)



write.table(gene_info.stat, file = paste0(gene_info.stat.file, ".tsv"), sep="\t", quote=F, col.names=NA)
